+++ src/scripts/defaults.ts
@@ -40,6 +40,7 @@ export const DEFAULT_CONFIG: Options = {
   silent: false,
   renderChoiceLimit: -1,
   maxItemCount: -1,
+  closeDropdownOnSelect: 'auto',
   singleModeForMultiSelect: false,
   addChoices: false,
   addItems: true,